mm/memory/hotplug: convert device bool to int to allow for more flags
Patch series "HMM (Heterogeneous Memory Management)", v18.


This patch (of 16):

When hotpluging memory we want more informations on the type of memory and
its properties.  Replace the device boolean flag by an int and define a
set of flags.

New property for device memory is an opt-in flag to allow page migration
from and to a ZONE_DEVICE.  Existing user of ZONE_DEVICE are not expecting
page migration to work for their pages.  New changes to page migration i
changing that and we now need a flag to explicitly opt-in page migration.

Link: http://lkml.kernel.org/r/1489680335-6594-2-git-send-email-jglisse@redhat.com
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Cc: Russell King <linux@armlinux.org.uk>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Martin Schwidefsky <schwidefsky@de.ibm.com>
Cc: Heiko Carstens <heiko.carstens@de.ibm.com>
Cc: Yoshinori Sato <ysato@users.sourceforge.jp>
Cc: Rich Felker <dalias@libc.org>
Cc: Chris Metcalf <cmetcalf@mellanox.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>