mm/ZONE_DEVICE/x86: add support for un-addressable device memory
It does not need much, just skip populating kernel linear mapping for
range of un-addressable device memory (it is pick so that there is no
physical memory resource overlapping it).  All the logic is in share mm
code.

Only support x86-64 as this feature doesn't make much sense with
constrained virtual address space of 32bits architecture.

Link: http://lkml.kernel.org/r/1489680335-6594-6-git-send-email-jglisse@redhat.com
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>