mm/migrate: allow migrate_vma() to alloc new page on empty entry
This allow caller of migrate_vma() to allocate new page for empty CPU page
table entry.  It only support anoymous memory and it won't allow new page
to be instance if userfaultfd is armed.

This is usefull to device driver that want to migrate a range of virtual
address and would rather allocate new memory than having to fault later
on.

Link: http://lkml.kernel.org/r/1489680335-6594-15-git-send-email-jglisse@redhat.com
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>