mm-page_alloc-split-smallest-stolen-page-in-fallback-fix
The lkp-robot reported a test case stuck on boot due to the patch [1]
which was due to endless loop in the modified __rmqueue().  It blindly
expected that move_freepages_block() will succeed, but that can fail due
to last pageblock pfn not belonging to the same zone as the fallback
candidate page.

This fix checks the result of move_freepages_block() and steals the single
candidate page if it fails, which was also effectively done before [1].

[1] mmotm: mm-page_alloc-split-smallest-stolen-page-in-fallback.patch

Link: http://lkml.kernel.org/r/59d71b35-d556-4fc9-ee2e-1574259282fd@suse.cz
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Stephen Rothwell <sfr@canb.auug.org.au>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>