zram: try harder to store user data on compression error
__zram_bvec_write() overwrites the existing compressed object only when it
successfully compresses a new one.  When zram_compress() fails, we
propagate the error, but never touch the old object, so we keep the old
data for that particular page and all reads that could hit that index
later will read stale data.  This can cause problems.  For example,

- Overwriting a file

     over-write block 1 aaa->xxx OK
     over-write block 2 bbb->yyy OK
     over-write block 3 ccc->zzz error << the block still has 'ccc' data.

- Now, reading the file

     read block 1 xxx OK
     read block 2 yyy OK
     read block 3 ccc OK   << we should not return OK here. Because
                              "xxxyyyccc" is not correct file.

As a solution, when we fail to compress a new page (S/W or H/W compressor
error, depending on particular setup) try to store the page uncompressed
(PAGE_SIZE-ed zs_pool object).  This, at least, allows us to return valid
and correct data:

     read block 1  xxx - OK, compressed
     read block 2  yyy - OK, compressed
     read block 3  zzz - OK, uncompressed (because compressing back-end
                                           returned an error)

Link: http://lkml.kernel.org/r/20170518021714.3229-1-sergey.senozhatsky@gmail.com
Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>