mm, THP, swap: Fix two issues in THP optimize patch
When changing the logic for cluster allocation for THP in get_swap_page(),
I made a mistake so that a normal swap slot may be allocated for a THP
instead of return with failure.  This is fixed in the patch.

And I found two likely/unlikely annotation is wrong in get_swap_pages(),
because that is slow path, I just removed the likely/unlikely annotation.

Link: http://lkml.kernel.org/r/87k25ed8zo.fsf@yhuang-dev.intel.com
Signed-off-by: "Huang, Ying" <ying.huang@intel.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>