binfmt_elf: document load_bias a little bit
Patch series "move mmap_area and PIE binaries away from the stack".

There are a few bugs causing the kernel to sometimes map PIE binaries and
the mmap_area where the stack is supposed to go.

This series fixes them for x86, ARM64, and PPC.  S390 seems to be ok.

This patch (of 6):

After I and another unnamed developer got confused by the subtraction of
vaddr in this branch of the code, followed by adding vaddr back in a
little bit later, for the third time, maybe it is time to document this
quirky bit of code.

Link: http://lkml.kernel.org/r/20170602152010.2064-2-riel@redhat.com
Signed-off-by: Rik van Riel <riel@redhat.com>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>
Cc: Daniel Micay <danielmicay@gmail.com>
Cc: Will Deacon <will.deacon@arm.com>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Florian Weimer <fweimer@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>