memcg: nano-optimize VM_BUG_ON in uncharge_page
Even though VM_BUG* is usually not compiled in there are systems which
enable CONFIG_DEBUG_VM by default.  VM_BUG_ON_PAGE in uncharge_page is not
very optimal for the normal case.  All pages should have counter==0 so
that is the first thing to check.  is_zone_device_page is nicely noop if
ZONE_DEVICE is not compiled in and finally HW poison pages should be least
probable.  So reorder the check to bail out as early as possible on normal
case.

Link: http://lkml.kernel.org/r/20170918063141.6coovns7cb45bfly@dhcp22.suse.cz
Signed-off-by: Michal Hocko <mhocko@suse.com>
Cc: Jérôme Glisse <jglisse@redhat.com>
Cc: Evgeny Baskakov <ebaskakov@nvidia.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>